@@ -25,7 +25,7 @@ apply plugin: 'com.android.application'
 apply from: "$flutterRoot/packages/flutter_tools/gradle/flutter.gradle"
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     lintOptions {
         disable 'InvalidPackage'
@@ -34,27 +34,20 @@ android {
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
         applicationId "com.szrlzz.qms"
-        minSdkVersion 16
-        targetSdkVersion 27
+        minSdkVersion 19
+        targetSdkVersion 28
         versionCode 2
         versionName 'QMS.0.2.0'
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        ndk {
-            // 设置支持的SO库架构
-            //abiFilters 'armeabi' //, 'x86', 'armeabi-v7a', 'x86_64', 'arm64-v8a'
-            abiFilters "armeabi", "armeabi-v7a", "x86", "mips", "arm64-v8a"
-        }
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     //签名配置
     signingConfigs {
         release {
-            release {
-                keyAlias 'ruilian'
-                keyPassword 'ruilian123456'
-                storeFile file('../rlzz_qms.jks')
-                storePassword 'ruilian123456'
-            }
+            keyAlias 'ruilian'
+            keyPassword 'ruilian123456'
+            storeFile file('../rlzz_qms.jks')
+            storePassword 'ruilian123456'
         }
     }
 
@@ -63,13 +56,26 @@ android {
             // TODO: Add your own signing config for the release build.
             // Signing with the debug keys for now, so `flutter run --release` works.
             signingConfig signingConfigs.release
+            /*ndk {
+                //选择要添加的对应 cpu 类型的 .so 库。
+                abiFilters  "x86"
+                // 还可以添加
+            }*/
+            //ndk { abiFilters "x86" }
+            ndk {
+                // 设置支持的SO库架构
+                //abiFilters 'armeabi' //, 'x86', 'armeabi-v7a', 'x86_64', 'arm64-v8a'
+                abiFilters  "armeabi-v7a"//, "x86", "mips", "arm64-v8a"
+            }
         }
+
     }
+
     compileOptions {
         sourceCompatibility = 1.8
         targetCompatibility = 1.8
     }
-    buildToolsVersion = '28.0.3'
+    //buildToolsVersion = '28.0.3'
 }
 
 flutter {
@@ -78,14 +84,17 @@ flutter {
 
 dependencies {
 
-    implementation 'com.android.support:support-v4:27.1.1'
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
     implementation 'com.github.PhilJay:MPAndroidChart:v3.0.2'
     implementation 'org.jetbrains:annotations:15.0'
+    implementation 'com.tencent.bugly:crashreport_upgrade:1.4.1'//其中latest.release指代最新版本号，也可以指定明确的版本号，例如1.2.0
+    implementation 'com.tencent.bugly:nativecrashreport:3.7.1' //其中latest.release指代最新版本号，也可以指定明确的版本号，例如2.2.0
+
 
 }
