@@ -1,4 +1,4 @@
-package com.szrlzz.qms.filepicker;
+package com.szrlzz.qms.plugin;
 
 import android.Manifest;
 import android.app.Activity;
@@ -7,8 +7,8 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.webkit.MimeTypeMap;
 
@@ -45,7 +45,7 @@ public class FilePickerPlugin implements MethodCallHandler {
         return;
     }
 
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), "file_picker");
+    final MethodChannel channel = new MethodChannel(registrar.messenger(), "com.szrlzz.qms/file_picker");
     channel.setMethodCallHandler(new FilePickerPlugin());
 
     instance = registrar;
