@@ -3,10 +3,10 @@ import 'dart:convert';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:connectivity/connectivity.dart';
 import 'package:qms/common/local/MySelfInfo.dart';
 import 'package:qms/common/net/LogUtils.dart';
 import 'package:qms/common/utils/CommonUtil.dart';
+import 'package:qms/common/utils/plugin/Connectivity.dart';
 import 'package:qms/common/utils/DeviceUtil.dart';
 import 'package:qms/common/config/Config.dart';
 import 'package:qms/common/net/Code.dart';
