@@ -4,10 +4,10 @@ import 'dart:typed_data';
 import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:path_provider/path_provider.dart';
 import 'package:qms/common/style/StringZh.dart';
 import 'package:qms/common/style/Styles.dart';
 import 'package:qms/common/utils/CommonUtil.dart';
+import 'package:qms/common/utils/plugin/PathProvider.dart';
 import 'package:qms/widget/ButtonWidget.dart';
 
 class SignaturePainterPage extends StatefulWidget {
