@@ -7,7 +7,7 @@ import 'package:qms/common/config/Config.dart';
 import 'package:qms/common/modal/TestOrder.dart';
 import 'package:qms/common/style/Styles.dart';
 import 'package:qms/common/utils/CommonUtil.dart';
-import 'package:qms/page/BadReasonRefPage.dart';
+import 'package:qms/page/RefPage.dart';
 import 'package:qms/page/SignPage.dart';
 import 'package:qms/widget/InputWidget.dart';
 import 'package:qms/widget/TextWidget.dart';
@@ -26,9 +26,9 @@ class TestOrderHeadInfoPage extends StatefulWidget {
 
   TestOrderHeadInfoPage({
     Key key,
-    this.testOrderInfo,
-    this.isAdd,
-    this.auditStatus,
+    @required this.testOrderInfo,
+    @required this.isAdd,
+    @required this.auditStatus,
   }) : super(key: key);
 
   @override
@@ -533,7 +533,7 @@ class TestOrderHeadInfoPageState extends State<TestOrderHeadInfoPage> {
         context: context, //BuildContext对象
         barrierDismissible: false,
         builder: (BuildContext context) {
-          return new BadReasonRefPage(
+          return new RefPage(url:Config.qmsApiUrl+Config.badReasonRefUrl,
             arcCode: widget.testOrderInfo.badReasonCode,
             okFun: (obj) {
               setState(() {
