@@ -4,17 +4,23 @@ import 'package:qms/common/style/Styles.dart';
 
 ///列表底部分页
 class ListPageWidget extends StatefulWidget {
-  final String page;
+  final int page;
+  final int size;
+  final int total;
+  final Function firstFun;
   final Function preFun;
   final Function nextFun;
-  final Function refreshFun;
+  final Function endFun;
 
   ListPageWidget({
     Key key,
     this.page,
+    this.size,
+    this.total,
+    this.firstFun,
     this.preFun,
     this.nextFun,
-    this.refreshFun,
+    this.endFun,
   }) : super(key: key);
 
   @override
@@ -27,6 +33,20 @@ class ListPageWidgetState extends State<ListPageWidget> {
     super.initState();
   }
 
+  int getEndPage() {
+    ///取余
+    int yu = widget.total % widget.size;
+
+    ///取整
+    int zheng = widget.total ~/ widget.size;
+
+    int newPage = zheng;
+    if (yu > 0) {
+      newPage = newPage + 1;
+    }
+    return newPage;
+  }
+
   @override
   Widget build(BuildContext context) {
     return new Container(
@@ -36,6 +56,21 @@ class ListPageWidgetState extends State<ListPageWidget> {
       child: new Row(
         mainAxisAlignment: MainAxisAlignment.end,
         children: <Widget>[
+          new GestureDetector(
+            onTap: () {
+              widget.firstFun();
+            },
+            child: new Container(
+              color: Colors.transparent,
+              alignment: Alignment.center,
+              width: 80.0,
+              height: 45.0,
+              child: new Text(StringZh.first_page,
+                  style: new TextStyle(
+                    fontSize: RLZZConstant.middleTextSize,
+                  )),
+            ),
+          ),
           new GestureDetector(
             onTap: () {
               widget.preFun();
@@ -53,11 +88,11 @@ class ListPageWidgetState extends State<ListPageWidget> {
           ),
           new Container(
             alignment: Alignment.center,
-            width: 20.0,
+            //width: 40.0,
             height: 45.0,
             margin: new EdgeInsets.only(right: 40.0, left: 40.0),
             child: new Text(
-              widget.page,
+              widget.page.toString() + '/' + getEndPage().toString()+'（' + widget.total.toString() + '）',
               style: new TextStyle(
                 fontSize: RLZZConstant.middleTextSize,
                 //color: RLZZColors.mainColor,
@@ -73,7 +108,7 @@ class ListPageWidgetState extends State<ListPageWidget> {
               alignment: Alignment.center,
               width: 80.0,
               height: 45.0,
-              margin: new EdgeInsets.only(right: 80.0),
+              //margin: new EdgeInsets.only(right: 80.0),
               child: new Text(
                 StringZh.next_page,
                 style: new TextStyle(
@@ -85,17 +120,21 @@ class ListPageWidgetState extends State<ListPageWidget> {
           ),
           new GestureDetector(
             onTap: () {
-              widget.refreshFun();
+              widget.endFun();
             },
             child: new Container(
               color: Colors.transparent,
               alignment: Alignment.center,
               width: 80.0,
               height: 45.0,
-              child: new Text(StringZh.refresh_page,
-                  style: new TextStyle(
-                    fontSize: RLZZConstant.middleTextSize,
-                  )),
+              //margin: new EdgeInsets.only(right: 80.0),
+              child: new Text(
+                StringZh.end_page,
+                style: new TextStyle(
+                  fontSize: RLZZConstant.middleTextSize,
+                  //color: RLZZColors.mainColor,
+                ),
+              ),
             ),
           ),
         ],
