@@ -4,14 +4,14 @@ import 'dart:io';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:qms/common/modal/Enclosure.dart';
 import 'package:qms/common/net/ApiUtil.dart';
 import 'package:qms/common/net/MethodChannelUtil.dart';
 import 'package:qms/common/style/StringZh.dart';
 import 'package:qms/common/style/Styles.dart';
 import 'package:qms/common/utils/CommonUtil.dart';
-import 'package:qms/common/utils/FilePicker.dart';
+import 'package:qms/common/utils/plugin/FilePicker.dart';
+import 'package:qms/common/utils/plugin/ImagePicker.dart';
 import 'package:qms/common/utils/WidgetUtil.dart';
 
 class UploadImageWidget extends StatefulWidget {
